search functionality: another step: adding the add feature to add the activity apart from using from the recommendations and search results
1) lifelogmain.go: since the add functionality is possible in two ways;
by form, and by url, I have added 2 specific add handlers instead of the
generic one
2) handlers.go: created separate handler functions for add-by-form
(handleActivityAddByForm) and add-by-URL (handleActivityAddByURL)
functionality. to generalize, moved the common code in the both handler
functions into a generic function (handleActivityAdd).
3) searchActivity.html: added the add form at the bottom of the
recommendations, so that user can directly add a activity
4) searchResults.html: removed the footer. (+) button is not required as
we are already in search page. also added the add form at the bottom of
the search results, so that user can directly add a activity if not
found in the search results
5) _ActivityList.html: since the add functionality is possible in two
ways; by form, and by url, I have specified the correct handler for add
in this form