added full text feature to the activity "search"; using appengine/search API
Successfully linked the datastore entries to the full-text-search-index
documents. the way it is done is tricky, how it’s made possible is;
whenever an activity is created/added, I take the activity name and add
to the index. to keep the activity names in index unique, I made the
index id same as the activity name (except that the spaces in activity
name are replaced with underscores, as spaces are not allowed in the
index id. this way could successfully maintain the unique activity names
in the index. Next, when the search is made, it is redirected to the
index, so that full text search feature can be used to retrieve the
records. the retrieved records in this case will be the activity names,
which can later be passed to data store query filter in case if the
complete activity log entry needs to be retrieved, or can be directly
used to list the activities based on the full test search results (which
is the case in the code at present).

Next: below are few of the pending tasks to make this perfect,
1) in helpers.FullTextSearchActivity(), the len of the slice ‘dst’ is
hardcoded to 10, as I’m not yet able to figure out how many records are
returned by the full text search query. there is a count function on the
full text search result object, but that is not helping, as it is just a
counter which is incremented when the .Next() used, but not the actual
count of the results. need some research in this area.
2) the full text search, to what I have explored, is not intelligent. if
there is a activity by name ‘sleep123’, and I search with keyword
‘sleep’, it is not retrieved as it looks like sleep123 not equal to
sleep. however, if the activity name was ‘sleep 123’, i.e. with a space
in between, the full text search would match. need to figure out if
there is a way to match more intelligently, like the Google search would
retrieve.
3) Need to add code in some where in the beginning of the application,
in order to reindex all the fields, as it looks like the index documents
are not staying after the application reboot. or may be I need to
research in order to find if there is a alternative setting with index
to persist. need to research.
4) some code cleanup and documentation is required to make the code loop
perfect, if not better :)